pyproject.toml
[tool.poetry]
name = "cutpanion"
version = "0.1.0"
description = ""
authors = ["zbednarke <zbednarke@gmail.com>"]
readme = "README.md"
packages = [{include = "cutpanion"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
pandas = "^1.5.1"
tqdm = "^4.64.1"
DateTime = "^4.7"
orjson = "^3.8.1"
matplotlib = "^3.6.2"
sqlalchemy = "^1.4.43"
psycopg2 = "^2.9.5"
django = "^4.1.4"
django-storages = "^1.13.2"
google-auth = "^2.15.0"
google-cloud-storage = "^2.7.0"
psycopg2-binary = "^2.9.5"
plotly = "^5.11.0"
gunicorn = "^20.1.0"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
 
[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
